Move changes to run_capsule
